Renamed module to not conflict with datasets variable. reflect new module name in init.py. this closes #26
