+++ openpolicedata/__init__.py
@@ -1,10 +1,10 @@
 import sys
 import os
 
-from .datasets import DataTypes
-from .datasets import TableTypes
-from .datasets import datasets
-from .datasets import get
+from ._datasets import DataTypes
+from ._datasets import TableTypes
+from ._datasets import datasets
+from ._datasets import get
 from .data import Source
